+++ packages/expo-secure-store/CHANGELOG.md
@@ -7,7 +7,7 @@
 ### 🎉 New features
 
 - Add `canUseDeviceCredentialsAuthentication` function. ([#41409](https://github.com/expo/expo/pull/41409) by [@JakubKorytko](https://github.com/JakubKorytko))
-- Add `enableDeviceFallback` option. ([#41409](https://github.com/expo/expo/pull/41409) by [@JakubKorytko](https://github.com/JakubKorytko))
+- Add `requireAuthentication: 'userPresence'` for device credential fallback (on Android, requires API 30+). ([#41409](https://github.com/expo/expo/pull/41409) by [@JakubKorytko](https://github.com/JakubKorytko))
 
 ### 🐛 Bug fixes